Added and fixed up licenses in all of the sources files.
In some cases I removed copyright notices attributing authorship to Tsong Chong
where they appeared by accident --- that is, in a couple of files where they had
been brought along when I made a copy of another file to serve as a template,
but which contained no actual code that Tsong Chong had wrote.

I also switched to the ISC license since it is more concise, and added some text
explaining the relationship of Illuminate to unit-- and exactly where code is
shared between them --- which is only in a few of the test files, and nowhere in
the library itself.

I considered stripping out his e-mail address since I know from experience that
it no longer works, but I changed my mind and decided to keep it in.